Raspberry Pi/Linux Macro Conflict
This may apply to Linux in general, but it certainly applies to the raspberry pi.  It appears that KEY_DOWN & KEY_UP are macros used by linux in input-event-codes.h and are conflicting with the same named variables in DOM Events.h Please note that this is a quick fix and is simply renaming those variables to DOM_KEY_DOWN & DOM_KEY_UP.  I have not dug very deep on this issue, so I am not sure if it breaks something, but it seems to resolve the issue.